Bluetooth: shell: Fix bsim xterm-native-shell
The xterm-native-shell snippet was broken since
CONFIG_UART_NATIVE_PTY_0_ON_STDINOUT was added, which redirected the
native PTY to stdin/stdout instead of the PTY xterm connected to. This
change disables it and does so explicitly with `=n`, even though that's
the default, in case the default changes. This restores output to the
xterm.

Its DT overlay also created two native PTYs instead of one, resulting
two xterms, only one of which is used, the other is blank. This change
removes overlay causing the second useless PTY.

The way the the PTY was being instantiated was also haphazardly done by
monkey-patching the compatible string on the UART nodes. This commit
changes the overlay to precisely create a new DT node for the PTY and
assign this node as chosen for zephyr,console and zephyr,shell-uart,
leaving the UART nodes alone in case they are used for something else.

Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>